working on file production
