+++ cggtts/src/lib.rs
@@ -174,34 +174,6 @@ use serde::{Deserialize, Serialize};
 /// Latest CGGTTS release : only version we truly support
 pub const CURRENT_RELEASE: &str = "2E";
 
-/*
-pub struct ITRF {
-    x: f64,
-    y: f64,
-    z: f64,
-    epoch: chrono::NaiveDateTime,
-}
-
-pub struct HelmertCoefs {
-    /// translation coefficients (cx, cy, cz) in meters
-    c: (f64, f64, f64),
-    /// scaling [ppb]
-    s: f64,
-    /// rotation matrix coefficients (rx, ry, rz) in arc second
-    r: (f64, f64, f64),
-}
-
-/// Transforms input 3D vector using Helmert transforms.
-/// Used to convert ITRF and other data
-pub fn helmert_transform (v: (f64,f64,f64), h: HelmertCoefs) -> (f64,f64,f64)  {
-    let (x, y, z) = v;
-    let (cx, cy, cz) = h.c;
-    let (rx, ry, rz) = h.z;
-    let x = cx + (1.0 * h.s*1E-9) * (x - rz*y + ry * z);
-    let y = cy + (1.0 * h.s*1E-9) * (rz*x + y -rx*z);
-    let z = cz + (1.0 * h.s*1E-9) * (-ry*x +rx*y + z);
-}*/
-
 #[derive(Clone, Copy, PartialEq, Debug, EnumString)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[derive(Default)]
@@ -212,7 +184,7 @@ pub enum Code {
     P1,
     P2,
     E1,
-    E5a,
+    E5,
     B1,
     B2,
 }
@@ -225,7 +197,7 @@ impl std::fmt::Display for Code {
             Code::P1 => fmt.write_str("P1"),
             Code::P2 => fmt.write_str("P2"),
             Code::E1 => fmt.write_str("E1"),
-            Code::E5a => fmt.write_str("E5a"),
+            Code::E5 => fmt.write_str("E5"),
             Code::B1 => fmt.write_str("B1"),
             Code::B2 => fmt.write_str("B2"),
         }
@@ -1004,129 +976,12 @@ impl std::fmt::Display for CGGTTS {
 #[cfg(test)]
 mod test {
     use super::*;
-    use std::path::Path;
     #[test]
     fn test_code() {
         assert_eq!(Code::default(), Code::C1);
         assert_eq!(Code::from_str("C2").unwrap(), Code::C2);
         assert_eq!(Code::from_str("P1").unwrap(), Code::P1);
         assert_eq!(Code::from_str("P2").unwrap(), Code::P2);
-        assert_eq!(Code::from_str("E5a").unwrap(), Code::E5a);
-    }
-    #[test]
-    fn gzsy8259_568() {
-        let path = Path::new(env!("CARGO_MANIFEST_DIR"))
-            .join("..")
-            .join("data")
-            .join("standard")
-            .join("GZSY8259.568");
-
-        let fullpath = path.to_string_lossy().to_string();
-        let cggtts = CGGTTS::from_file(&fullpath);
-        assert_eq!(cggtts.is_ok(), true);
-
-        let cggtts = cggtts.unwrap();
-
-        assert_eq!(
-            cggtts.release_date,
-            Epoch::from_gregorian_utc_at_midnight(2014, 02, 20),
-        );
-
-        assert_eq!(
-            cggtts.rcvr,
-            Some(
-                Rcvr::default()
-                    .manufacturer("GORGYTIMING")
-                    .receiver("SYREF25")
-                    .serial_number("18259999")
-                    .year(2018)
-                    .release("v00")
-            ),
-        );
-
-        assert_eq!(cggtts.station, "SY82");
-        assert_eq!(cggtts.nb_channels, 12);
-        assert_eq!(cggtts.ims, None);
-        assert_eq!(
-            cggtts.reference_time,
-            ReferenceTime::Custom(String::from("REF(SY82)"))
-        );
-        assert_eq!(cggtts.reference_frame, Some(String::from("ITRF")));
-        assert!((cggtts.apc_coordinates.x - 4314143.824).abs() < 1E-6);
-        assert!((cggtts.apc_coordinates.y - 452633.241).abs() < 1E-6);
-        assert!((cggtts.apc_coordinates.z - 4660711.385).abs() < 1E-6);
-        assert_eq!(cggtts.comments, None);
-        assert_eq!(cggtts.tracks.len(), 32);
-
-        let first = cggtts.tracks.first();
-        //assert_eq!(cggtts.delay.value(), 0.0);
-        assert_eq!(first.is_some(), true);
-        let first = first.unwrap();
-        assert_eq!(
-            first.sv,
-            SV {
-                constellation: Constellation::GPS,
-                prn: 99,
-            }
-        );
-
-        assert_eq!(cggtts.filename(), String::from("GSSY1859.568"));
-
-        let tracks: Vec<_> = cggtts.tracks().collect();
-        assert_eq!(tracks.len(), 32);
-
-        let _dumped = cggtts.to_string();
-        let _compare = std::fs::read_to_string(
-            &(env!("CARGO_MANIFEST_DIR").to_owned() + "/../data/standard/GZSY8259.568"),
-        )
-        .unwrap();
-    }
-    #[test]
-    fn rzsy8257_000() {
-        let path = Path::new(env!("CARGO_MANIFEST_DIR"))
-            .join("..")
-            .join("data")
-            .join("advanced")
-            .join("RZSY8257.000");
-
-        let fullpath = path.to_string_lossy().to_string();
-        let cggtts = CGGTTS::from_file(&fullpath);
-        assert_eq!(cggtts.is_ok(), true);
-
-        let cggtts = cggtts.unwrap();
-        assert!(cggtts.rcvr.is_none());
-        assert!(cggtts.ims.is_none());
-        assert_eq!(
-            cggtts.apc_coordinates,
-            Coordinates {
-                x: 4027881.79,
-                y: 306998.67,
-                z: 4919499.36,
-            }
-        );
-        assert!(cggtts.comments.is_none());
-        assert_eq!(cggtts.station, "ABC");
-        assert_eq!(cggtts.nb_channels, 12);
-
-        assert_eq!(cggtts.delay.rf_cable_delay, 237.0);
-        assert_eq!(cggtts.delay.ref_delay, 149.6);
-        assert_eq!(cggtts.delay.delays.len(), 2);
-        assert_eq!(cggtts.delay.delays[0], (Code::C1, Delay::Internal(53.9)));
-
-        let total = cggtts.delay.total_delay(Code::C1);
-        assert_eq!(total.is_some(), true);
-        assert_eq!(total.unwrap(), 53.9 + 237.0 + 149.6);
-
-        assert_eq!(cggtts.delay.delays[1], (Code::C2, Delay::Internal(49.8)));
-        let total = cggtts.delay.total_delay(Code::C2);
-        assert_eq!(total.is_some(), true);
-        assert_eq!(total.unwrap(), 49.8 + 237.0 + 149.6);
-
-        let cal_id = cggtts.delay.cal_id.clone();
-        assert_eq!(cal_id.is_some(), true);
-        assert_eq!(cal_id.unwrap(), String::from("1nnn-yyyy"));
-
-        let tracks: Vec<_> = cggtts.tracks().collect();
-        assert_eq!(tracks.len(), 4);
+        assert_eq!(Code::from_str("E5").unwrap(), Code::E5);
     }
 }
