+++ cggtts/src/tests/parser.rs
@@ -1,9 +1,12 @@
 mod test {
-    use crate::tests::toolkit::{cmp_dut_model, random_name};
-    use crate::CGGTTS;
+    use crate::{
+        prelude::{Constellation, Epoch, Rcvr, ReferenceTime, CGGTTS, SV},
+        tests::toolkit::{cmp_dut_model, random_name},
+        Code, Coordinates, Delay,
+    };
     use std::fs::File;
     use std::io::Write;
-    use std::path::PathBuf;
+    use std::path::{Path, PathBuf};
     #[test]
     fn single_frequency_files() {
         let resources = PathBuf::new()
@@ -35,6 +38,16 @@ mod test {
             let mut fd = File::create(&filename).unwrap();
             write!(fd, "{}", cggtts).unwrap();
 
+            // parse back
+            let parsed = CGGTTS::from_file(&filename);
+            assert!(
+                parsed.is_ok(),
+                "failed to parse back generated file: {}",
+                parsed.err().unwrap()
+            );
+
+            cmp_dut_model(&parsed.unwrap(), &cggtts);
+
             // remove generated file
             let _ = std::fs::remove_file(&filename);
         }
@@ -45,6 +58,7 @@ mod test {
             .join(env!("CARGO_MANIFEST_DIR"))
             .join("../data")
             .join("advanced");
+
         for entry in std::fs::read_dir(resources).unwrap() {
             let entry = entry.unwrap();
             let path = entry.path();
@@ -62,5 +76,122 @@ mod test {
                 cggtts.err()
             );
         }
+    }
+    #[test]
+    fn gzsy8259_568() {
+        let path = Path::new(env!("CARGO_MANIFEST_DIR"))
+            .join("..")
+            .join("data")
+            .join("standard")
+            .join("GZSY8259.568");
+
+        let fullpath = path.to_string_lossy().to_string();
+        let cggtts = CGGTTS::from_file(&fullpath);
+        assert_eq!(cggtts.is_ok(), true);
+
+        let cggtts = cggtts.unwrap();
+
+        assert_eq!(
+            cggtts.release_date,
+            Epoch::from_gregorian_utc_at_midnight(2014, 02, 20),
+        );
+
+        assert_eq!(
+            cggtts.rcvr,
+            Some(
+                Rcvr::default()
+                    .manufacturer("GORGYTIMING")
+                    .receiver("SYREF25")
+                    .serial_number("18259999")
+                    .year(2018)
+                    .release("v00")
+            ),
+        );
+
+        assert_eq!(cggtts.station, "SY82");
+        assert_eq!(cggtts.nb_channels, 12);
+        assert_eq!(cggtts.ims, None);
+        assert_eq!(
+            cggtts.reference_time,
+            ReferenceTime::Custom(String::from("REF(SY82)"))
+        );
+
+        assert_eq!(cggtts.reference_frame, Some(String::from("ITRF")));
+        assert!((cggtts.apc_coordinates.x - 4314143.824).abs() < 1E-6);
+        assert!((cggtts.apc_coordinates.y - 452633.241).abs() < 1E-6);
+        assert!((cggtts.apc_coordinates.z - 4660711.385).abs() < 1E-6);
+        assert_eq!(cggtts.comments, None);
+        assert_eq!(cggtts.tracks.len(), 32);
+
+        let first = cggtts.tracks.first();
+        //assert_eq!(cggtts.delay.value(), 0.0);
+        assert_eq!(first.is_some(), true);
+        let first = first.unwrap();
+        assert_eq!(
+            first.sv,
+            SV {
+                constellation: Constellation::GPS,
+                prn: 99,
+            }
+        );
+
+        assert_eq!(cggtts.filename(), String::from("GSSY1859.568"));
+
+        let tracks: Vec<_> = cggtts.tracks().collect();
+        assert_eq!(tracks.len(), 32);
+
+        let _dumped = cggtts.to_string();
+        let _compare = std::fs::read_to_string(
+            &(env!("CARGO_MANIFEST_DIR").to_owned() + "/../data/standard/GZSY8259.568"),
+        )
+        .unwrap();
+    }
+    #[test]
+    fn rzsy8257_000() {
+        let path = Path::new(env!("CARGO_MANIFEST_DIR"))
+            .join("..")
+            .join("data")
+            .join("advanced")
+            .join("RZSY8257.000");
+
+        let fullpath = path.to_string_lossy().to_string();
+        let cggtts = CGGTTS::from_file(&fullpath);
+        assert_eq!(cggtts.is_ok(), true);
+
+        let cggtts = cggtts.unwrap();
+        assert!(cggtts.rcvr.is_none());
+        assert!(cggtts.ims.is_none());
+        assert_eq!(
+            cggtts.apc_coordinates,
+            Coordinates {
+                x: 4027881.79,
+                y: 306998.67,
+                z: 4919499.36,
+            }
+        );
+        assert!(cggtts.comments.is_none());
+        assert_eq!(cggtts.station, "ABC");
+        assert_eq!(cggtts.nb_channels, 12);
+
+        assert_eq!(cggtts.delay.rf_cable_delay, 237.0);
+        assert_eq!(cggtts.delay.ref_delay, 149.6);
+        assert_eq!(cggtts.delay.delays.len(), 2);
+        assert_eq!(cggtts.delay.delays[0], (Code::C1, Delay::Internal(53.9)));
+
+        let total = cggtts.delay.total_delay(Code::C1);
+        assert_eq!(total.is_some(), true);
+        assert_eq!(total.unwrap(), 53.9 + 237.0 + 149.6);
+
+        assert_eq!(cggtts.delay.delays[1], (Code::C2, Delay::Internal(49.8)));
+        let total = cggtts.delay.total_delay(Code::C2);
+        assert_eq!(total.is_some(), true);
+        assert_eq!(total.unwrap(), 49.8 + 237.0 + 149.6);
+
+        let cal_id = cggtts.delay.cal_id.clone();
+        assert_eq!(cal_id.is_some(), true);
+        assert_eq!(cal_id.unwrap(), String::from("1nnn-yyyy"));
+
+        let tracks: Vec<_> = cggtts.tracks().collect();
+        assert_eq!(tracks.len(), 4);
     }
 }
